Customize response timeout for HTTP requests
This patch introduces a customized HTTP transport settings for all
crawler's activity (GET, HEAD) allowing to define a response timeout.

The timeout is the time the HTTP clients waits for response header
after completing writting a request to the socket. If the timeout
is exceeded, the client stops waiting and returns an error.

The timeout value can be defined for each profile watcher differently.
It is set by configuration parameter "timeout" in every profile
configuration.

Change-Id: I0697d46506a402fa2c1b3bf86558e32d7bc76f7e
Signed-off-by: Lukasz Wojciechowski <l.wojciechow@partner.samsung.com>