Add schema_elem structure implementation and tests
schema_elem structure describes a single database schema entity like
table or trigger.

create() method uses given transaction to create and verify existense
of table or trigger in the database. If there is no such element yet,
then it is created. If there already exist an element with such type
and name it's schema is verified and error is returned when schema
doesn't match.

Change-Id: I7876532991a29d80caaf9686bd4db045832e08ce
Signed-off-by: Lukasz Wojciechowski <l.wojciechow@partner.samsung.com>